add illustrator support
