+++ .changeset/dull-seals-help.md
@@ -1,0 +1,6 @@
+---
+"opendesign": minor
+"@opendesign/universal": minor
+---
+
+added experimental adobe illustrator support, currently node.js-only
