+++ packages/opendesign-universal/src/node.ts
@@ -1,0 +1,1 @@
+export { importIllustratorFile } from "./octopus-file/importer-ai.js";
